\name{bvseqs}
\alias{bvseqs}
\docType{data}
\title{
BV reference set.
}
\description{
Tree-derived pairwise distances and taxonomic assignments among 16S rRNA
sequences representing bacteria represented in the vaginal mucosa.
}
\usage{data(bvseqs)}
\format{\preformatted{
  The format is:
List of 3
 $ dmat    : num [1:448, 1:448] 0 0.0494 0.0968 0.1002 0.1606 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:448] "S001098970" "S000859776" "S000539896" "S001352901" ...
  .. ..$ : chr [1:448] "S001098970" "S000859776" "S000539896" "S001352901" ...
 $ groupTab:'data.frame':	448 obs. of  12 variables:
  ..$ superkingdom    : chr [1:448] "2" "2" "2" "2" ...
  ..$ superphylum     : chr [1:448] NA NA NA NA ...
  ..$ phylum          : chr [1:448] "1224" "1224" "1224" "1224" ...
  ..$ class           : chr [1:448] "1236" "1236" "1236" "1236" ...
  ..$ subclass        : chr [1:448] NA NA NA NA ...
  ..$ order           : chr [1:448] "72274" "72274" "72274" "72274" ...
  ..$ suborder        : chr [1:448] NA NA NA NA ...
  ..$ family          : chr [1:448] "468" "468" "468" "468" ...
  ..$ genus           : chr [1:448] "469" "469" "469" "469" ...
  ..$ species_group   : chr [1:448] NA NA NA NA ...
  ..$ species_subgroup: chr [1:448] NA NA NA NA ...
  ..$ species         : chr [1:448] "470" "470" "471" "470" ...
 $ taxNames: Named chr [1:212] "Actinomyces urogenitalis" "Lactobacillus jensenii" "Proteobacteria" "Gammaproteobacteria" ...
  ..- attr(*, "names")= chr [1:212] "103621" "109790" "1224" "1236" ...
}}
\details{
(Describe creation of this data set)
}
\source{
Sequences were assembled from both the RDP 16S rRNA database and from
the laboratory of Dr. David Fredricks.
}
\references{
RDP url here.
}
\examples{
data(bvseqs)
## maybe str(bvseqs) ; plot(bvseqs) ...
}
\keyword{datasets}
